@@ -3,12 +3,12 @@ description: An example project for the flutter popup dialog.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.0.0-dev.68.0 <3.0.0"
+  sdk: ">=2.7.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  cupertino_icons: ^0.1.2
+  cupertino_icons: 1.0.0
   rflutter_alert:
     path: ../
 
