@@ -5,7 +5,7 @@ authors: [Ratel <rflutter@ratel.com.tr>]
 homepage: https://github.com/RatelHub/rflutter_alert
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.7.0 <3.0.0"
 
 dependencies:
   flutter:
