@@ -19,142 +19,142 @@ import 'dialog_button.dart';
 ///
 /// You must call the "show()" method to view the alert class you have defined.
 class Alert {
-    final String id;
-    final BuildContext context;
-    final AlertType type;
-    final AlertStyle style;
-    final Image image;
-    final String title;
-    final String desc;
-    final Widget content;
-    final List<DialogButton> buttons;
-    final ButtonsContainer buttonsContainer;
-    final Function closeFunction;
-    final Icon closeIcon;
-    final bool onWillPopActive;
+  final String id;
+  final BuildContext context;
+  final AlertType type;
+  final AlertStyle style;
+  final Image image;
+  final String title;
+  final String desc;
+  final Widget content;
+  final List<DialogButton> buttons;
+  final ButtonsContainer buttonsContainer;
+  final Function closeFunction;
+  final Icon closeIcon;
+  final bool onWillPopActive;
 
 
-    /// Alert constructor
-    ///
-    /// [context], [title] are required.
-    Alert({
-        @required this.context,
-        this.id,
-        this.type,
-        this.style = const AlertStyle(),
-        this.image,
-        @required this.title,
-        this.desc,
-        this.content,
-        this.buttons,
-        this.buttonsContainer = ButtonsContainer.row,
-        this.closeFunction,
-        this.closeIcon,
-        this.onWillPopActive = false,
-    });
+  /// Alert constructor
+  ///
+  /// [context], [title] are required.
+  Alert({
+    @required this.context,
+    this.id,
+    this.type,
+    this.style = const AlertStyle(),
+    this.image,
+    @required this.title,
+    this.desc,
+    this.content,
+    this.buttons,
+    this.buttonsContainer = ButtonsContainer.row,
+    this.closeFunction,
+    this.closeIcon,
+    this.onWillPopActive = false,
+  });
 
-    /// Displays defined alert window
-    Future<bool> show() async {
-        return await showGeneralDialog(
-            context: context,
-            pageBuilder: (BuildContext buildContext, Animation<double> animation,
-                Animation<double> secondaryAnimation) {
-                return _buildDialog();
-            },
-            barrierDismissible: style.isOverlayTapDismiss,
-            barrierLabel: MaterialLocalizations
-                .of(context)
-                .modalBarrierDismissLabel,
-            barrierColor: style.overlayColor,
-            useRootNavigator: true,
-            transitionDuration: style.animationDuration,
-            transitionBuilder: (BuildContext context,
-                Animation<double> animation,
-                Animation<double> secondaryAnimation,
-                Widget child,) =>
-                _showAnimation(animation, secondaryAnimation, child),
-        );
-    }
+  /// Displays defined alert window
+  Future<bool> show() async {
+    return await showGeneralDialog(
+      context: context,
+      pageBuilder: (BuildContext buildContext, Animation<double> animation,
+          Animation<double> secondaryAnimation) {
+        return _buildDialog();
+      },
+      barrierDismissible: style.isOverlayTapDismiss,
+      barrierLabel: MaterialLocalizations
+          .of(context)
+          .modalBarrierDismissLabel,
+      barrierColor: style.overlayColor,
+      useRootNavigator: true,
+      transitionDuration: style.animationDuration,
+      transitionBuilder: (BuildContext context,
+          Animation<double> animation,
+          Animation<double> secondaryAnimation,
+          Widget child,) =>
+          _showAnimation(animation, secondaryAnimation, child),
+    );
+  }
 
-    // Will be added in next version.
-    Future<void> dismiss() async {
-        Navigator.of(context, rootNavigator: true).pop();
-    }
+  // Will be added in next version.
+  Future<void> dismiss() async {
+    Navigator.of(context, rootNavigator: true).pop();
+  }
 
-    // Alert dialog content widget
-    Widget _buildDialog() {
-        final Widget _child = Center(
-            child: ConstrainedBox(
-                constraints: style.constraints ?? BoxConstraints.expand(width: double.infinity, height: double.infinity),
+  // Alert dialog content widget
+  Widget _buildDialog() {
+    final Widget _child = Center(
+      child: ConstrainedBox(
+        constraints: style.constraints ?? BoxConstraints.expand(
+            width: double.infinity, height: double.infinity),
+        child: Center(
+          child: AlertDialog(
+              key: Key(id),
+              backgroundColor: style.backgroundColor ?? Theme
+                  .of(context)
+                  .dialogBackgroundColor,
+              shape: style.alertBorder ?? _defaultShape(),
+              titlePadding: const EdgeInsets.all(0.0),
+              title: Container(
                 child: Center(
-                    child: AlertDialog(
-                        key: Key(id),
-                        backgroundColor: style.backgroundColor ?? Theme
-                            .of(context)
-                            .dialogBackgroundColor,
-                        shape: style.alertBorder ?? _defaultShape(),
-                        titlePadding: const EdgeInsets.all(0.0),
-                        title: Container(
-                            child: Center(
-                                child: Column(
-                                    mainAxisAlignment: MainAxisAlignment.center,
-                                    children: <Widget>[
-                                        _getCloseButton(),
-                                        Padding(
-                                            padding: EdgeInsets.fromLTRB(
-                                                20, (style.isCloseButton ? 0 : 10), 20, 0),
-                                            child: Column(
-                                                children: <Widget>[
-                                                    _getImage(),
-                                                    SizedBox(
-                                                        height: 15,
-                                                    ),
-                                                    Text(
-                                                        title,
-                                                        style: style.titleStyle,
-                                                        textAlign: style.titleTextAlign,
-                                                    ),
-                                                    SizedBox(
-                                                        height: desc == null ? 5 : 10,
-                                                    ),
-                                                    desc == null
-                                                        ? Container()
-                                                        : Text(
-                                                        desc,
-                                                        style: style.descStyle,
-                                                        textAlign: style.descTextAlign,
-                                                    ),
-                                                    content == null ? Container() : content,
-                                                ],
-                                            ),
-                                        )
-                                    ],
-                                ),
+                  child: Column(
+                    mainAxisAlignment: MainAxisAlignment.center,
+                    children: <Widget>[
+                      _getCloseButton(),
+                      Padding(
+                        padding: EdgeInsets.fromLTRB(
+                            20, (style.isCloseButton ? 0 : 10), 20, 0),
+                        child: Column(
+                          children: <Widget>[
+                            _getImage(),
+                            SizedBox(
+                              height: 15,
                             ),
+                            Text(
+                              title,
+                              style: style.titleStyle,
+                              textAlign: style.titleTextAlign,
+                            ),
+                            SizedBox(
+                              height: desc == null ? 5 : 10,
+                            ),
+                            desc == null
+                                ? Container()
+                                : Text(
+                              desc,
+                              style: style.descStyle,
+                              textAlign: style.descTextAlign,
+                            ),
+                            content == null ? Container() : content,
+                          ],
                         ),
-                        contentPadding: style.buttonAreaPadding,
-                        content: buttonsContainer == ButtonsContainer.row ? Row(
-                            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                            children: _getButtons(),
-                        ) : Container(
-                            alignment: Alignment.center,
-                            height: 150,
-                            child: Column(
-                                crossAxisAlignment: CrossAxisAlignment.center,
-                                mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                                children: _getButtons(),
-                            ),
-                        )
-                    ),
+                      )
+                    ],
+                  ),
                 ),
-            ),
+              ),
+              contentPadding: style.buttonAreaPadding,
+              content: buttonsContainer == ButtonsContainer.row ? Row(
+                mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+                children: _getButtons(),
+              ) : Container(
+                alignment: Alignment.center,
+                height: 150,
+                child: Column(
+                  crossAxisAlignment: CrossAxisAlignment.center,
+                  mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+                  children: _getButtons(),
+                ),
+              )
           ),
         ),
       ),
     );
     return onWillPopActive
-        ? WillPopScope(onWillPop: () async => false, child: _child)
-        : _child;
+    ? WillPopScope(onWillPop: () async => false, child: _child
+    )
+    :
+    _child;
   }
 
 
@@ -162,142 +162,146 @@ class Alert {
   Widget _getCloseButton() {
     return style.isCloseButton
         ? Padding(
-            padding: const EdgeInsets.fromLTRB(0, 10, 10, 0),
-            child:
-            GestureDetector(
-              onTap: (){ 
-                Navigator.pop(context);
-                if(closeFunction != null){
-                  closeFunction();
-                }
-              },
-              child: Container(
-              alignment: FractionalOffset.topRight,
+      padding: const EdgeInsets.fromLTRB(0, 10, 10, 0),
+      child:
+      GestureDetector(
+        onTap: () {
+          Navigator.pop(context);
+          if (closeFunction != null) {
+            closeFunction();
+          }
+        },
+        child: Container(
+          alignment: FractionalOffset.topRight,
 
-              child: this.closeIcon != null ?
-              Container(child: this.closeIcon)
-                  : Container(
-                width: 20,
-                height: 20,
-                decoration: BoxDecoration(
-                  image: DecorationImage(
-                    image: AssetImage(
-                      '$kImagePath/close.png',
-                      package: 'rflutter_alert',
-                    ),
-                  ),
+          child: this.closeIcon != null ?
+          Container(child: this.closeIcon)
+              : Container(
+            width: 20,
+            height: 20,
+            decoration: BoxDecoration(
+              image: DecorationImage(
+                image: AssetImage(
+                  '$kImagePath/close.png',
+                  package: 'rflutter_alert',
                 ),
               ),
             ),
-            ),
-          )
+          ),
+        ),
+      ),
+    )
         : Container();
   }
-  
-    // Returns defined buttons. Default: Cancel Button
-    List<Widget> _getButtons() {
-        List<Widget> expandedButtons = [];
-        if (style.isButtonVisible) {
-            if (buttons != null) {
-                buttons.forEach(
-                        (button) {
-                        var buttonWidget = Padding(
-                            padding: const EdgeInsets.only(left: 2, right: 2),
-                            child: button,
-                        );
-                        if (button.width != null && buttons.length == 1) {
-                            expandedButtons.add(buttonWidget);
-                        } else {
-                            expandedButtons.add(Expanded(
-                                child: buttonWidget,
-                            ));
-                        }
-                    },
-                );
+
+  // Returns defined buttons. Default: Cancel Button
+  List<Widget> _getButtons() {
+    List<Widget> expandedButtons = [];
+    if (style.isButtonVisible) {
+      if (buttons != null) {
+        buttons.forEach(
+              (button) {
+            var buttonWidget = Padding(
+              padding: const EdgeInsets.only(left: 2, right: 2),
+              child: button,
+            );
+            if (button.width != null && buttons.length == 1) {
+              expandedButtons.add(buttonWidget);
             } else {
-                expandedButtons.add(
-                    Expanded(
-                        child: DialogButton(
-                            child: Text(
-                                "CANCEL",
-                                style: TextStyle(
-                                    color: Colors.white, fontSize: 20),
-                            ),
-                            onPressed: () => Navigator.pop(context),
-                        ),
-                    ),
-                );
+              expandedButtons.add(Expanded(
+                child: buttonWidget,
+              ));
             }
-        }
-    }
-
-    // Returns alert default border style
-    ShapeBorder _defaultShape() {
-        return RoundedRectangleBorder(
-            borderRadius: BorderRadius.circular(10.0),
-            side: BorderSide(
-                color: Colors.blueGrey,
-            ),
+          },
         );
+      } else {
+        expandedButtons.add(
+          Expanded(
+            child: DialogButton(
+              child: Text(
+                "CANCEL",
+                style: TextStyle(
+                    color: Colors.white, fontSize: 20),
+              ),
+              onPressed: () => Navigator.pop(context),
+            ),
+          ),
+        );
+      }
     }
+    return expandedButtons;
+  }
 
-    // Returns alert image for icon
-    Widget _getImage() {
-        Widget response = image ?? Container();
-        switch (type) {
-            case AlertType.success:
-                response = Image.asset(
-                    '$kImagePath/icon_success.png',
-                    package: 'rflutter_alert',
-                );
-                break;
-            case AlertType.error:
-                response = Image.asset(
-                    '$kImagePath/icon_error.png',
-                    package: 'rflutter_alert',
-                );
-                break;
-            case AlertType.info:
-                response = Image.asset(
-                    '$kImagePath/icon_info.png',
-                    package: 'rflutter_alert',
-                );
-                break;
-            case AlertType.warning:
-                response = Image.asset(
-                    '$kImagePath/icon_warning.png',
-                    package: 'rflutter_alert',
-                );
-                break;
-            case AlertType.none:
-                response = Container();
-                break;
-        }
-        return response;
-    }
+  // Returns alert default border style
+  ShapeBorder _defaultShape() {
+    return RoundedRectangleBorder(
+      borderRadius: BorderRadius.circular(10.0),
+      side: BorderSide(
+        color: Colors.blueGrey,
+      ),
+    );
+  }
 
-    // Shows alert with selected animation
-    _showAnimation(animation, secondaryAnimation, child) {
-        switch (style.animationType) {
-            case AnimationType.fromRight:
-                return AnimationTransition.fromRight(animation, secondaryAnimation, child);
-                break;
-            case AnimationType.fromLeft:
-                return AnimationTransition.fromLeft(animation, secondaryAnimation, child);
-                break;
-            case AnimationType.fromBottom:
-                return AnimationTransition.fromBottom(
-                    animation, secondaryAnimation, child);
-                break;
-            case AnimationType.grow:
-                return AnimationTransition.grow(animation, secondaryAnimation, child);
-                break;
-            case AnimationType.shrink:
-                return AnimationTransition.shrink(animation, secondaryAnimation, child);
-                break;
-            case AnimationType.fromTop:
-                return AnimationTransition.fromTop(animation, secondaryAnimation, child);
-                break;
-        }
+  // Returns alert image for icon
+  Widget _getImage() {
+    Widget response = image ?? Container();
+    switch (type) {
+      case AlertType.success:
+        response = Image.asset(
+          '$kImagePath/icon_success.png',
+          package: 'rflutter_alert',
+        );
+        break;
+      case AlertType.error:
+        response = Image.asset(
+          '$kImagePath/icon_error.png',
+          package: 'rflutter_alert',
+        );
+        break;
+      case AlertType.info:
+        response = Image.asset(
+          '$kImagePath/icon_info.png',
+          package: 'rflutter_alert',
+        );
+        break;
+      case AlertType.warning:
+        response = Image.asset(
+          '$kImagePath/icon_warning.png',
+          package: 'rflutter_alert',
+        );
+        break;
+      case AlertType.none:
+        response = Container();
+        break;
     }
+    return response;
+  }
+
+  // Shows alert with selected animation
+  _showAnimation(animation, secondaryAnimation, child) {
+    switch (style.animationType) {
+      case AnimationType.fromRight:
+        return AnimationTransition.fromRight(
+            animation, secondaryAnimation, child);
+        break;
+      case AnimationType.fromLeft:
+        return AnimationTransition.fromLeft(
+            animation, secondaryAnimation, child);
+        break;
+      case AnimationType.fromBottom:
+        return AnimationTransition.fromBottom(
+            animation, secondaryAnimation, child);
+        break;
+      case AnimationType.grow:
+        return AnimationTransition.grow(animation, secondaryAnimation, child);
+        break;
+      case AnimationType.shrink:
+        return AnimationTransition.shrink(animation, secondaryAnimation, child);
+        break;
+      case AnimationType.fromTop:
+        return AnimationTransition.fromTop(
+            animation, secondaryAnimation, child);
+        break;
+    }
+  }
 }
