@@ -9,97 +9,96 @@
 import 'package:flutter/material.dart';
 
 /// Predefined functions for transition animations
-/// 
+///
 /// Exp: AnimationTransition.fromRight(animation, secondaryAnimation, child);
 class AnimationTransition {
+  /// Slide animation, from right to left (SlideTransition)
+  static fromRight(Animation<double> animation,
+      Animation<double> secondaryAnimation, Widget child) {
+    return SlideTransition(
+      position: Tween<Offset>(
+        begin: const Offset(1.0, 0.0),
+        end: Offset.zero,
+      ).animate(animation),
+      child: child,
+    );
+  }
 
-    /// Slide animation, from right to left (SlideTransition)
-    static fromRight(Animation<double> animation,
-        Animation<double> secondaryAnimation, Widget child) {
-        return SlideTransition(
-            position: Tween<Offset>(
-                begin: const Offset(1.0, 0.0),
-                end: Offset.zero,
-            ).animate(animation),
-            child: child,
-        );
-    }
+  /// Slide animation, from left to right (SlideTransition)
+  static fromLeft(Animation<double> animation,
+      Animation<double> secondaryAnimation, Widget child) {
+    return SlideTransition(
+      position: Tween<Offset>(
+        begin: const Offset(-1.0, 0.0),
+        end: Offset.zero,
+      ).animate(animation),
+      child: child,
+    );
+  }
 
-    /// Slide animation, from left to right (SlideTransition)
-    static fromLeft(Animation<double> animation,
-        Animation<double> secondaryAnimation, Widget child) {
-        return SlideTransition(
-            position: Tween<Offset>(
-                begin: const Offset(-1.0, 0.0),
-                end: Offset.zero,
-            ).animate(animation),
-            child: child,
-        );
-    }
+  /// Slide animation, from top to bottom (SlideTransition)
+  static fromTop(Animation<double> animation,
+      Animation<double> secondaryAnimation, Widget child) {
+    return SlideTransition(
+      position: Tween<Offset>(
+        begin: const Offset(0.0, -1.0),
+        end: Offset.zero,
+      ).animate(animation),
+      child: child,
+    );
+  }
 
-    /// Slide animation, from top to bottom (SlideTransition)
-    static fromTop(Animation<double> animation,
-        Animation<double> secondaryAnimation, Widget child) {
-        return SlideTransition(
-            position: Tween<Offset>(
-                begin: const Offset(0.0, -1.0),
-                end: Offset.zero,
-            ).animate(animation),
-            child: child,
-        );
-    }
+  /// Slide animation, from top to bottom (SlideTransition)
+  static fromBottom(Animation<double> animation,
+      Animation<double> secondaryAnimation, Widget child) {
+    return SlideTransition(
+      position: Tween<Offset>(
+        begin: const Offset(0.0, 1.0),
+        end: Offset.zero,
+      ).animate(animation),
+      child: child,
+    );
+  }
 
-    /// Slide animation, from top to bottom (SlideTransition)
-    static fromBottom(Animation<double> animation,
-        Animation<double> secondaryAnimation, Widget child) {
-        return SlideTransition(
-            position: Tween<Offset>(
-                begin: const Offset(0.0, 1.0),
-                end: Offset.zero,
-            ).animate(animation),
-            child: child,
-        );
-    }
+  /// Scale animation, from in to out (ScaleTransition)
+  static grow(Animation<double> animation, Animation<double> secondaryAnimation,
+      Widget child) {
+    return ScaleTransition(
+      scale: Tween<double>(
+        begin: 0.0,
+        end: 1.0,
+      ).animate(
+        CurvedAnimation(
+          parent: animation,
+          curve: Interval(
+            0.00,
+            0.50,
+            curve: Curves.linear,
+          ),
+        ),
+      ),
+      child: child,
+    );
+  }
 
-    /// Scale animation, from in to out (ScaleTransition)
-    static grow(Animation<double> animation,
-        Animation<double> secondaryAnimation, Widget child) {
-        return ScaleTransition(
-            scale: Tween<double>(
-                begin: 0.0,
-                end: 1.0,
-            ).animate(
-                CurvedAnimation(
-                    parent: animation,
-                    curve: Interval(
-                        0.00,
-                        0.50,
-                        curve: Curves.linear,
-                    ),
-                ),
-            ),
-            child: child,
-        );
-    }
-
-    /// Scale animation, from out to in (ScaleTransition)
-    static shrink(Animation<double> animation,
-        Animation<double> secondaryAnimation, Widget child) {
-        return ScaleTransition(
-            scale: Tween<double>(
-                begin: 1.2,
-                end: 1.0,
-            ).animate(
-                CurvedAnimation(
-                    parent: animation,
-                    curve: Interval(
-                        0.50,
-                        1.00,
-                        curve: Curves.linear,
-                    ),
-                ),
-            ),
-            child: child,
-        );
-    }
+  /// Scale animation, from out to in (ScaleTransition)
+  static shrink(Animation<double> animation,
+      Animation<double> secondaryAnimation, Widget child) {
+    return ScaleTransition(
+      scale: Tween<double>(
+        begin: 1.2,
+        end: 1.0,
+      ).animate(
+        CurvedAnimation(
+          parent: animation,
+          curve: Interval(
+            0.50,
+            1.00,
+            curve: Curves.linear,
+          ),
+        ),
+      ),
+      child: child,
+    );
+  }
 }
